Move some resources to the correct package
