+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/ee/injection/mappedname/web.xml
