+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/ee/injection/resource/basic/web.xml
