+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/ejb/mdb/ejb-jar.xml
