+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/pojo/test/a-jboss-beans.xml
