+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/pojo/test/b-jboss-beans.xml
