+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/sar/application.xml
