+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/sar/jboss-service-with-application-xml.xml
